requirements.txt
streamlit
firebase-admin
Pillow
google-generativeai
python-dotenv
